refactor(platform): create everruns-platform crate and move organization/principal models out of core
Carve the backend identity family out of everruns-core into a new
workspace crate, everruns-platform. This first slice moves the two
durable aggregate entity structs that the "Done when" bar names:

- everruns_core::Organization -> everruns_platform::Organization
- everruns_core::Principal    -> everruns_platform::Principal

Cross-cutting value types that core's runtime, permissions layer, and
domain models embed stay in everruns-core and are re-exported from the
platform crate for a unified consumer surface:

- OrgRole / OrgMembership and the multitenancy constants
  (DEFAULT_ORG_ID, DEFAULT_ORG_PUBLIC_ID, ANONYMOUS_USER_*, org
  public-id helpers) — used by core's permissions/auth layer.
- PrincipalKind / PrincipalStatus / PrincipalSummary — embedded in
  Session/App/SessionSchedule and the agent-identity lifecycle.

Moving those out would require either a core -> platform edge
(forbidden) or relocating permissions/Session/App/SessionSchedule
(out of scope), so they remain in core. See the PR for the follow-up.

The dependency direction is strictly platform -> core; core gains no
edge to platform. A new dependency-assertion test
(crates/core/tests/no_platform_dependency.rs) fails the core build if
that edge is ever introduced. Serialized shapes and OpenAPI output are
unchanged — the structs move verbatim with identical derives.

Publish wiring updated for the new crate: workspace members and
[workspace.dependencies], scripts/sync-publish-pin-versions.py
(INNER_PINS + WORKSPACE_PIN_DEPS), and the publish-crates workflow
(dependency_versions map + CRATES publish order, after everruns-core).

Fixes EVE-837
