+++ crates/core/src/principal.rs
@@ -1,12 +1,17 @@
-// Principal domain types
+// Principal domain value types.
 //
 // Design Decision:
 // - Principals are org-scoped durable owners/executors with structured metadata.
 // - Ownership resolves through principal lineage to exactly one human user.
 // - Execution provenance remains separate from ownership; principals only model
 //   who can own durable entities and who an unattended flow can act as.
+//
+// EVE-837: the `Principal` aggregate entity moved to the `everruns-platform`
+// crate. The value types below stay in core because they are embedded by core
+// domain models: `PrincipalSummary` is a field of `Session`/`App`/
+// `SessionSchedule`, `PrincipalKind` backs that summary, and `PrincipalStatus`
+// is shared with the agent-identity lifecycle.
 
-use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
@@ -89,39 +94,4 @@ pub struct PrincipalSummary {
     pub subject_id: Option<Uuid>,
     #[serde(default, skip_serializing_if = "serde_json::Value::is_null")]
     pub metadata: serde_json::Value,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[cfg_attr(feature = "openapi", derive(ToSchema))]
-pub struct Principal {
-    #[cfg_attr(feature = "openapi", schema(value_type = String, example = "principal_01933b5a000070008000000000000001"))]
-    pub id: PrincipalId,
-    pub organization_id: String,
-    pub kind: PrincipalKind,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub subject_id: Option<Uuid>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub parent_principal_id: Option<PrincipalId>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub resolved_user_id: Option<Uuid>,
-    #[serde(default, skip_serializing_if = "serde_json::Value::is_null")]
-    pub metadata: serde_json::Value,
-    pub status: PrincipalStatus,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub archived_at: Option<DateTime<Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deleted_at: Option<DateTime<Utc>>,
-}
-
-impl Principal {
-    pub fn summary(&self) -> PrincipalSummary {
-        PrincipalSummary {
-            id: self.id,
-            kind: self.kind.clone(),
-            subject_id: self.subject_id,
-            metadata: self.metadata.clone(),
-        }
-    }
 }
