+++ crates/server/Cargo.toml
@@ -86,6 +86,8 @@ tar.workspace = true
 bashkit = { version = "0.16.0", features = ["scripted_tool", "jq"] }
 
 everruns-core = { path = "../core", features = ["embedded-platform-docs", "openapi", "sqlx", "tree-sitter-outlines", "ui-capabilities"] }
+# Backend identity entities (Organization, Principal) — moved out of core in EVE-837.
+everruns-platform = { path = "../platform", features = ["openapi"] }
 # Observability exporters (Braintrust, OpenTelemetry event listeners). The OTLP
 # init these rely on stays in core's `telemetry` feature (on by default).
 everruns-observability = { workspace = true }
